revert the changes to BPE tokenizer and move changes to SmolVLMTokenizer
