+++ src/main/scala/com/johnsnowlabs/nlp/annotators/tokenizer/bpe/SmolVLMTokenizer.scala
@@ -16,7 +16,7 @@
 
 package com.johnsnowlabs.nlp.annotators.tokenizer.bpe
 
-import com.johnsnowlabs.nlp.annotators.common.IndexedToken
+import com.johnsnowlabs.nlp.annotators.common.{IndexedToken, Sentence}
 
 import java.nio.charset.Charset
 import scala.collection.mutable.ListBuffer
@@ -107,6 +107,83 @@ class SmolVLMTokenizer(
   //    }
   //    decoded.toString().replaceAll(decoderVocab(29871), " ").trim()
   //  }
+
+  override def tokenize(sentence: Sentence): Array[IndexedToken] = {
+    var originalText = sentence.content
+    if (originalText.trim.isEmpty) return Array.empty[IndexedToken]
+
+    if (padWithSequenceTokens) {
+      originalText = sentencePadding._1 + originalText + sentencePadding._2
+    }
+
+    var textList: ListBuffer[String] = ListBuffer(originalText)
+
+    for (specialToken <- specialTokens.allTokens) {
+      val currentTokenContent = specialToken.content
+      if (currentTokenContent.nonEmpty) { // Avoid issues with empty special tokens
+        val regex = java.util.regex.Pattern.quote(currentTokenContent).r
+        val newList = ListBuffer[String]()
+        for (segment <- textList) {
+          if (specialTokens.contains(segment)) {
+            // Already identified as a special token in a previous iteration
+            newList += segment
+          } else {
+            // Split this segment by the current special token, preserving the token
+            var lastIndex = 0
+            regex.findAllMatchIn(segment).foreach { m =>
+              if (m.start > lastIndex) {
+                newList += segment.substring(lastIndex, m.start)
+              }
+              newList += m.matched // Add the special token itself
+              lastIndex = m.end
+            }
+            if (lastIndex < segment.length) {
+              newList += segment.substring(lastIndex)
+            }
+          }
+        }
+        textList = newList.filter(_.nonEmpty) // Update list and remove potential empty strings
+      }
+    }
+
+    // Adjust index calculations to account for potential padding
+    val textWithPadding = originalText // Use the potentially padded text for indexing
+    val indexOffsetAdjustment =
+      if (padWithSequenceTokens) sentence.start - sentencePadding._1.length else sentence.start
+
+    var currentIndex = 0
+    val result = mutable.ArrayBuffer[IndexedToken]()
+    for (subText <- textList) {
+      // Find the start index based on the text *with* padding if applied
+      val subTextIndexInPadded = textWithPadding.indexOf(subText, currentIndex)
+      if (subTextIndexInPadded == -1) {
+        // This case should ideally not happen if splitting logic is correct, but handle defensively
+        // Log warning or error? For now, skip.
+        println(
+          s"Warning: Could not find segment '$subText' starting from index $currentIndex in text '$textWithPadding'")
+        // Attempt to recover by searching from the beginning, though this might be wrong
+        currentIndex = textWithPadding.indexOf(subText, 0) + subText.length
+      } else {
+        currentIndex = subTextIndexInPadded
+
+        // Calculate original index relative to the *unpadded* sentence start
+        val originalStartIndex = currentIndex + indexOffsetAdjustment
+        val originalEndIndex = originalStartIndex + subText.length - 1
+
+        if (!specialTokens.contains(subText)) {
+          // Pass the original start index as the offset for sub-tokenization
+          val splitSubText: Array[IndexedToken] = tokenizeSubText(subText, originalStartIndex)
+          result.append(splitSubText: _*)
+        } else {
+          // It's a special token, create IndexedToken directly using original indices
+          result.append(IndexedToken(subText, begin = originalStartIndex, end = originalEndIndex))
+        }
+        // Move currentIndex forward in the padded text
+        currentIndex += subText.length
+      }
+    }
+    result.toArray
+  }
   def decodeTokens(tokens: Array[Int]): String = {
     val text = tokens
       .map(token => decoderVocab(token))
